+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog
 1.0.2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix ``concurrent.Future`` exceptions not being propagated.
 
 
 1.0.1 (2015-12-17)